surface height restructure
